Add xscom_ok() and lpc_ok() to check XSCOM and LPC usability
This primarily checks whether the caller already holds the corresponding
locks to avoid re-entrancy in some of the deep error path such as when
XSCOM itself triggers an error log. It will be extended in the case of
LPC to also handle known HW error states.

We use them to avoid queuing/polling in the BT driver and to discard
characters in the UART driver.

Note: This will not normally involve a loss of log to the UART as the
UART driver is also protected by the console suspend mechanism. So
this is a safety mechanism only.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>